Move Core files into folders
