+++ src/NaCl.Core/Base/ChaChaCore/ChaCha20CoreIntrinsics.cs
@@ -4,7 +4,7 @@ namespace NaCl.Core.Base;
 using System;
 using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
-
+using NaCl.Core.Base.ChaChaCore;
 
 internal class ChaCha20CoreIntrinsics : IChaCha20Core
 {
